Fixed pitch. Issues with static

+++ src/main/java/legend/core/audio/MidiChannel.java
@@ -3,5 +3,5 @@ package legend.core.audio;
 public interface MidiChannel {
   double getVolume();
   int getPan();
-  int getPitchBend();
+  double getPitchBend();
 }
